LayoutTests/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/json-module/parse-error-expected.txt

CONSOLE MESSAGE: SyntaxError: JSON Parse error: Expected '}'

PASS JSON modules: parse error
